requirements.txt
cmudict==1.0.15
pandas==2.1.3
praat-textgrids==1.3.1
tqdm
nltk==3.8.1
transformers==4.35.0
jiwer==3.0.3
librosa==0.10.1
fasttext==0.9.2

# Better if you used conda:
# conda install pytorch==2.0.1 torchaudio==2.0.2 pytorch-cuda=11.8 -c pytorch -c nvidia
torch==2.0.1
torchaudio==2.0.2

# Optional
# jupyterlab
# matplotlib
# seaborn
# scipy

# textless @ git+https://github.com/facebookresearch/textlesslib@ba33d66
fairseq @ git+https://github.com/pytorch/fairseq@dd106d9